LRAC-14746 Simplify
