+++ modules/apps/analytics/analytics-batch-export-import-impl/src/main/java/com/liferay/analytics/batch/exportimport/internal/dispatch/executor/DXPEntityAnalyticsExportDispatchTaskExecutor.java
@@ -5,13 +5,26 @@
 
 package com.liferay.analytics.batch.exportimport.internal.dispatch.executor;
 
+import com.liferay.analytics.batch.exportimport.manager.AnalyticsBatchExportImportManager;
 import com.liferay.analytics.dxp.entity.rest.dto.v1_0.DXPEntity;
 import com.liferay.analytics.settings.configuration.AnalyticsConfigurationRegistry;
+import com.liferay.dispatch.executor.BaseDispatchTaskExecutor;
 import com.liferay.dispatch.executor.DispatchTaskExecutor;
 import com.liferay.dispatch.executor.DispatchTaskExecutorOutput;
+import com.liferay.dispatch.executor.DispatchTaskStatus;
+import com.liferay.dispatch.model.DispatchLog;
 import com.liferay.dispatch.model.DispatchTrigger;
+import com.liferay.dispatch.service.DispatchLogLocalService;
+import com.liferay.petra.function.UnsafeConsumer;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 
 import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
 
 import org.osgi.service.component.annotations.Component;
@@ -28,7 +41,7 @@ import org.osgi.service.component.annotations.Reference;
 	service = DispatchTaskExecutor.class
 )
 public class DXPEntityAnalyticsExportDispatchTaskExecutor
-	extends BaseAnalyticsDXPEntityExportDispatchTaskExecutor {
+	extends BaseDispatchTaskExecutor {
 
 	public static final String KEY = "export-analytics-dxp-entities";
 
@@ -38,7 +51,7 @@ public class DXPEntityAnalyticsExportDispatchTaskExecutor
 			DispatchTaskExecutorOutput dispatchTaskExecutorOutput)
 		throws Exception {
 
-		if (!shouldExport(dispatchTrigger.getCompanyId())) {
+		if (!_analyticsConfigurationRegistry.isActive()) {
 			return;
 		}
 
@@ -57,15 +70,62 @@ public class DXPEntityAnalyticsExportDispatchTaskExecutor
 		return KEY;
 	}
 
-	@Override
-	protected String getBatchEngineExportTaskItemDelegateName() {
-		return "analytics-dxp-entities";
+	protected UnsafeConsumer<String, Exception> getNotificationUnsafeConsumer(
+		long dispatchTriggerId,
+		DispatchTaskExecutorOutput dispatchTaskExecutorOutput) {
+
+		DispatchLog dispatchLog =
+			dispatchLogLocalService.fetchLatestDispatchLog(
+				dispatchTriggerId, DispatchTaskStatus.IN_PROGRESS);
+
+		return message -> updateDispatchLog(
+			dispatchLog.getDispatchLogId(), dispatchTaskExecutorOutput,
+			message);
 	}
 
-	@Override
-	protected boolean shouldExport(long companyId) {
-		return _analyticsConfigurationRegistry.isActive();
+	protected Date getResourceLastModifiedDate(long dispatchTriggerId) {
+		DispatchLog latestSuccessfulDispatchLog =
+			dispatchLogLocalService.fetchLatestDispatchLog(
+				dispatchTriggerId, DispatchTaskStatus.SUCCESSFUL);
+
+		if (latestSuccessfulDispatchLog == null) {
+			return null;
+		}
+
+		return latestSuccessfulDispatchLog.getEndDate();
 	}
+
+	protected void updateDispatchLog(
+			long dispatchLogId,
+			DispatchTaskExecutorOutput dispatchTaskExecutorOutput,
+			String message)
+		throws PortalException {
+
+		StringBundler sb = new StringBundler(5);
+
+		if (dispatchTaskExecutorOutput.getOutput() != null) {
+			sb.append(dispatchTaskExecutorOutput.getOutput());
+		}
+
+		sb.append(_dateFormat.format(new Date()));
+		sb.append(StringPool.SPACE);
+		sb.append(message);
+		sb.append(StringPool.NEW_LINE);
+
+		dispatchTaskExecutorOutput.setOutput(sb.toString());
+
+		dispatchLogLocalService.updateDispatchLog(
+			dispatchLogId, new Date(), dispatchTaskExecutorOutput.getError(),
+			dispatchTaskExecutorOutput.getOutput(),
+			DispatchTaskStatus.IN_PROGRESS);
+	}
+
+	@Reference
+	protected AnalyticsBatchExportImportManager
+		analyticsBatchExportImportManager;
+
+	@Reference
+	protected DispatchLogLocalService dispatchLogLocalService;
 
 	private static final List<String> _batchEngineExportTaskItemDelegateNames =
 		Arrays.asList(
@@ -78,6 +138,8 @@ public class DXPEntityAnalyticsExportDispatchTaskExecutor
 			"organization-analytics-dxp-entities",
 			"role-analytics-dxp-entities", "team-analytics-dxp-entities",
 			"user-analytics-dxp-entities", "user-group-analytics-dxp-entities");
+	private static final DateFormat _dateFormat = new SimpleDateFormat(
+		"yyyy-MM-dd'T'HH:mm:ss.SSSZ");
 
 	@Reference
 	private AnalyticsConfigurationRegistry _analyticsConfigurationRegistry;
